Update HTML tag and Webring alt text
The HTML tag within the index.html file has been enhanced with the inclusion of a language attribute. The webring logo also was missing alt text, so it has been added. These modifications are aimed at bettering accessibility by aiding assistive technologies and search engines to understand the language used for the content on the webpage more effectively.